worker/requirements.txt
psycopg2==2.9.3
python-dotenv==0.20.0
redis==4.2.2